Changing handling of SB API - Auto lookup WS connection handle using Receiver AP MAC
Using aruba_conn_id which is the value of the (supposed unique) req.headers['sec-websocket-key'] to uniquely identify an incoming WS connection.

Created handle_json_test_msg() to test JSON strings (apHealthUpdate msg) using wscat. Associate the aruba_conn_id with the AP mac by creating an array of APs with the conn_id as one of the values inside it.

the /test_api endpoint was created to check if SB msg is sent to the correct WS connection. We take the recv ap MAC from the incoming SB msg and check that in the aruba_ws_clients array to find the matching AP MAC, and hence the connection id. Iterate thru the ws.clients object and find the desired connection using the conn_id.